arreglo modificacioines en las vistas al incluir el navbar , arreglo problema del weight de la letra en los títulos, traduzco nombres de componentes en español a inglés
