+++ src/app/(header)/ongoing-delivery/page.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import OngoingDelivery from "components/OngoingDelivery";
+const RepartoEnCursoPage: React.FC = () => {
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-[#3d1df3]">
+      <OngoingDelivery />
+    </div>
+  );
+};
+
+export default RepartoEnCursoPage;
