--- src/app/(header)/reparto-en-curso/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import RepartoEnCurso from 'components/RepartoEnCurso'
-const RepartoEnCursoPage: React.FC = () => {
-	return (
-		<div className="min-h-screen flex items-center justify-center bg-[#3d1df3]">
-			<RepartoEnCurso />
-		</div>
-	)
-}
-
-export default RepartoEnCursoPage
