+++ src/app/layout.tsx
@@ -1,27 +1,27 @@
-import React from 'react'
-import type { Metadata } from 'next'
-import { Poppins } from 'next/font/google'
-import './globals.css'
+import React from "react";
+import type { Metadata } from "next";
+import { Poppins } from "next/font/google";
+import "./globals.css";
 
 const poppins = Poppins({
-	weight: '400',
-	subsets: ['latin'],
-	variable: '--poppins-default',
-})
+  weight: ["400", "700"],
+  subsets: ["latin"],
+  variable: "--poppins-default",
+});
 
 export const metadata: Metadata = {
-	title: 'Create Next App',
-	description: 'Generated by create next app',
-}
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
 
 export default function RootLayout({
-	children,
+  children,
 }: {
-	children: React.ReactNode;
+  children: React.ReactNode;
 }) {
-	return (
-		<html lang="en">
-			<body className={poppins.className}>{children}</body>
-		</html>
-	)
+  return (
+    <html lang="en">
+      <body className={poppins.className}>{children}</body>
+    </html>
+  );
 }
